+++ README.md
@@ -1,7 +1,15 @@
-# Descreption
+Welcome to My Ngram 
+
+My_ngram project
 
 # Task
+To create a function which counts number of elements on terminal
 
-# Usage
+# Description
+I use this project nesting for loops and use if for print number of the elements in string array  on terminal
 
 # Instalition
+first: write make on terminal and call ./my_ngram
+
+# Usage
+You may check it via writing any word after ./my_ngram
